* Moves validation into the bloc instead of UI. * Moves generateUsername into the state mapper instead of repository.
