+++ lib/v2/screens/sign_up/viewmodels/states/create_username_state.dart
@@ -54,4 +54,6 @@ class CreateUsernameState extends Equatable {
   }
 
   bool get isUsernameValid => !username.isNullOrEmpty && pageState == PageState.success;
+
+  bool get isNextButtonActive => isUsernameValid;
 }
